requirements.txt
# Основные зависимости проекта GGame
# Добавляйте сюда все необходимые пакеты

# Django и веб-фреймворк
Django==5.1.3
djangorestframework==3.15.2

# Асинхронность для Telegram Bot API
aiohttp==3.10.10
python-telegram-bot==21.6

# База данных
psycopg2-binary==2.9.10  # для PostgreSQL
sqlite3  # встроен в Python (для разработки)

# Дополнительные утилиты
python-dotenv==1.0.1  # для переменных окружения
requests==2.32.3

# Для разработки
django-debug-toolbar==4.4.6
